# webui/src/api/openapi.json
{
  "components": {
    "schemas": {
...
      "BilibiliStatus": {
        "properties": {
          "expired": {
            "default": false,
            "title": "Expired",
            "type": "boolean"
          },
          "face": {
            "default": "",
            "title": "Face",
            "type": "string"
          },
          "level": {
            "default": 0,
            "title": "Level",
            "type": "integer"
          },
          "logged_in": {
            "title": "Logged In",
            "type": "boolean"
          },
          "uid": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "string"
              }
            ],
            "default": 0,
            "title": "Uid"
          },
          "uname": {
            "default": "",
            "title": "Uname",
            "type": "string"
          },
          "vip_status": {
            "default": 0,
            "title": "Vip Status",
            "type": "integer"
          }
        },
        "required": [
          "logged_in"
        ],
        "title": "BilibiliStatus",
        "type": "object"
      },
...
      "CookieLoginRequest": {
        "properties": {
          "text": {
            "maxLength": 8192,
            "minLength": 1,
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "text"
        ],
        "title": "CookieLoginRequest",
        "type": "object"
      },
...
      "EpisodeList": {
        "description": "`episode` 是 TaskInfo 那套自由字典，原样回传给创建任务的接口即可",
        "properties": {
          "episodes": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Episodes",
            "type": "array"
          }
        },
        "required": [
          "episodes"
        ],
        "title": "EpisodeList",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ParseRequest": {
        "properties": {
          "pn": {
            "default": 1,
            "maximum": 10000.0,
            "minimum": 1.0,
            "title": "Pn",
            "type": "integer"
          },
          "url": {
            "maxLength": 2048,
            "minLength": 1,
            "title": "Url",
            "type": "string"
          }
        },
        "required": [
          "url"
        ],
        "title": "ParseRequest",
        "type": "object"
      },
...
      "QRCodeStatus": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "integer"
          },
          "message": {
            "default": "",
            "title": "Message",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "code",
          "status"
        ],
        "title": "QRCodeStatus",
        "type": "object"
      },
...
      "SMSSendRequest": {
        "properties": {
          "challenge": {
            "maxLength": 256,
            "minLength": 1,
            "title": "Challenge",
            "type": "string"
          },
          "cid": {
            "maxLength": 8,
            "minLength": 1,
            "title": "Cid",
            "type": "string"
          },
          "seccode": {
            "maxLength": 512,
            "minLength": 1,
            "title": "Seccode",
            "type": "string"
          },
          "tel": {
            "maxLength": 32,
            "minLength": 1,
            "title": "Tel",
            "type": "string"
          },
          "token": {
            "maxLength": 256,
            "minLength": 1,
            "title": "Token",
            "type": "string"
          },
          "validate": {
            "maxLength": 512,
            "minLength": 1,
            "title": "Validate",
            "type": "string"
          }
        },
        "required": [
          "cid",
          "tel",
          "token",
          "challenge",
          "validate",
          "seccode"
        ],
        "title": "SMSSendRequest",
        "type": "object"
      },
      "SMSSendResult": {
        "description": "captcha_key 是**发短信的返回值**，登录时才用，别与极验的三件套搞混",
        "properties": {
          "captcha_key": {
            "title": "Captcha Key",
            "type": "string"
          }
        },
        "required": [
          "captcha_key"
        ],
        "title": "SMSSendResult",
        "type": "object"
      },
      "SMSVerifyRequest": {
        "properties": {
          "captcha_key": {
            "maxLength": 512,
            "minLength": 1,
            "title": "Captcha Key",
            "type": "string"
          },
          "cid": {
            "maxLength": 8,
            "minLength": 1,
            "title": "Cid",
            "type": "string"
          },
          "code": {
            "maxLength": 16,
            "minLength": 1,
            "title": "Code",
            "type": "string"
          },
          "tel": {
            "maxLength": 32,
            "minLength": 1,
            "title": "Tel",
            "type": "string"
          }
        },
        "required": [
          "cid",
          "tel",
          "code",
          "captcha_key"
        ],
        "title": "SMSVerifyRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Bili23 Downloader WebUI",
    "version": "2.15.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/login/cookie": {
      "post": {
        "description": "用粘贴的 Cookie 登录。验证不通过会回滚，不会把无效 Cookie 留在 client 上",
        "operationId": "login_with_cookie_api_login_cookie_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CookieLoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BilibiliStatus"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Login With Cookie",
        "tags": [
          "login"
        ]
      }
    },
...
    "/api/login/qrcode/poll": {
      "get": {
        "description": "查一次扫码状态\n\n状态为 `success` 时 Cookie 已经写进配置了（在 `parse_poll` 里就地完成），\n前端接着调 `/api/login/status?refresh=true` 取用户信息即可",
        "operationId": "poll_qrcode_api_login_qrcode_poll_get",
        "parameters": [
          {
            "in": "query",
            "name": "key",
            "required": true,
            "schema": {
              "maxLength": 256,
              "minLength": 1,
              "title": "Key",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QRCodeStatus"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Poll Qrcode",
        "tags": [
          "login"
        ]
      }
    },
...
    "/api/login/sms/send": {
      "post": {
        "description": "发送验证码短信。返回的 captcha_key 登录时要用",
        "operationId": "sms_send_api_login_sms_send_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SMSSendRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SMSSendResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Sms Send",
        "tags": [
          "login"
        ]
      }
    },
    "/api/login/sms/verify": {
      "post": {
        "description": "用收到的验证码完成登录",
        "operationId": "sms_verify_api_login_sms_verify_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SMSVerifyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BilibiliStatus"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Sms Verify",
        "tags": [
          "login"
        ]
      }
    },
    "/api/login/status": {
      "get": {
        "description": "当前的 B 站登录态\n\n`refresh = True` 时去问一次 nav 接口（顺带刷新 wbi 签名用的 key）；\n否则只报配置里存的状态，不发请求 —— 前端轮询这个接口时不该每次都打一次 B 站",
        "operationId": "login_status_api_login_status_get",
        "parameters": [
          {
            "in": "query",
            "name": "refresh",
            "required": false,
            "schema": {
              "default": false,
              "title": "Refresh",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BilibiliStatus"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Login Status",
        "tags": [
          "login"
        ]
      }
    },
    "/api/parse": {
      "post": {
        "description": "解析一个链接，返回剧集树\n\n树的叶子节点带 `episode` 字段 —— 那正是创建下载任务时要交回来的东西，\n前端原样回传即可，不必自己拼",
        "operationId": "parse_url_api_parse_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ParseRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Parse Url",
        "tags": [
          "parse"
        ]
      }
    },
    "/api/parse/episodes": {
      "post": {
        "description": "从解析树里把叶子摘出来\n\n前端完全可以自己遍历，这里提供一份是为了**让摘取规则只有一处**：\n树节点（`is_node`）不算下载项，这条规则改了两边就会对不上",
        "operationId": "extract_episodes_api_parse_episodes_post",
        "parameters": [
          {
            "in": "query",
            "name": "only_checked",
            "required": false,
            "schema": {
              "default": true,
              "title": "Only Checked",
              "type": "boolean"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "title": "Tree",
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EpisodeList"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Extract Episodes",
        "tags": [
          "parse"
        ]
      }
    },
...
  }
}